requirements.txt
Django==5.1.3
django-crispy-forms==2.3
crispy-bootstrap4==2024.10
pillow==11.0.0